refactor: first service split
